Update Ruby from 1.9.3 to 2.1.2.
Used Ruby version MRI 1.9.3 is stale and it's officially recommended to
switch to one of 2.x branches because of security and support. Also some
feature and performance improvements (most singificant is GC) was made in
latest versions of MRI interpreter was made.
So, Ruby update simplify and improve development and production usage process.